+++ CHANGELOG.md
@@ -8,6 +8,12 @@ and this project adheres to
 
 ## [Unreleased]
 
+## Added
+
+- Added some grace to namespace configuration to allow for
+  `jupiterone.bamboohr.com` and `https://jupiterone.bamboohr.com` as well as
+  `jupiterone`
+
 ## 0.3.3 - 2020-11-30
 
 - Add the `supervisor` property to employee entities